Create billsplitter.py
First input is for how many people are participating in the bill splitting: For example 5.
Second input is for adding every person including you which are participating in the bill splitting.
Third is the option to choose if someone gets not to play the bill - to be lucky. The bill will be split accordingly to the choice of having a lucky person or not.
In the end a dictionary is printed out with a name and the amount they have to pay.